requirements.txt
asgiref==3.5.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.10
distlib==0.3.8
Django==3.2.9
django-appconf==1.0.5
django-bleach==2.0.0
django-ckeditor==6.4.0
django-environ==0.8.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-wkhtmltopdf==3.3.0
filelock==3.13.1
idna==3.3
packaging==21.3
pdfkit==1.0.0
pilkit==2.0
Pillow==11.3.0
platformdirs==4.2.0
pyparsing==3.0.9
pytz==2021.3
razorpay==1.2.0
requests==2.27.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
webencodings==0.5.1
